ada: Add contracts to Ada.Strings.Unbounded library
This patch adds contracts to the conversions between
Unbounded_String and String, the Element function and the
equality between two Unbounded_String, or between
Unbounded_String and String.
This patch also disallows the use of a function in SPARK, because
it returns an uninitialized Unbounded_String.

gcc/ada/

	* libgnat/a-strunb.ads, libgnat/a-strunb__shared.ads
	(To_Unbounded_String): Add postcondition. Add aspect SPARK_Mode
	Off on the version that takes a Natural as parameter.
	(To_String): Complete postcondition.
	(Set_Unbounded_String): Add postcondition.
	(Element): Likewise.
	("="): Likewise.

Tested on x86_64-pc-linux-gnu, committed on master.