ada: Fix missing finalization in library-unit instance spec
This fixes the missing finalization of objects declared in the spec of
package instances that are library units (and only them, i.e. not all
library-level package instances) when the instances have a package body.

The finalization is done when there is no package body, and supporting
this case precisely broke the other case because of a thinko or a typo.

This also requires a small adjustment to the routine writing ALI files.

gcc/ada/

	* exp_ch7.adb (Build_Finalizer): Reverse the test comparing the
	instantiation and declaration nodes of a package instance, and
	therefore bail out only when they are equal.  Adjust comments.
	(Expand_N_Package_Declaration): Do not clear the Finalizer field.
	* lib-writ.adb: Add with and use clauses for Sem_Util.
	(Write_Unit_Information): Look at unit nodes to find finalizers.
	* sem_ch12.adb (Analyze_Package_Instantiation): Beef up the comment
	about the rewriting of the instantiation node into a declaration.

Tested on x86_64-pc-linux-gnu, committed on master.