ada: Fix bogus error on predicated limited record declared in protected type
This happens when the limited record is initialized with a function call
because of a couple of issues: incorrect tree sharing when building the
predicate check and too late freezing for a compiler-generated subtype.

It turns out that building the predicate check manually is redundant here,
since predicate checks are automatically generated during the expansion of
assignment statements, and the late freezing can be easily fixed.

gcc/ada/

	* exp_ch3.adb (Build_Record_Init_Proc.Build_Assignment): Do not
	manually generate a predicate check.  Call Unqualify before doing
	pattern matching on the expression.
	* sem_ch3.adb (Analyze_Object_Declaration): Also freeze the actual
	subtype when it is built in the definite case.

Tested on x86_64-pc-linux-gnu, committed on master.