Dump if a pattern fails after having printed applying it
While trying to understand how to use the ! operand for match
patterns, I noticed that the debug dumps would print out applying
a pattern but nothing when it was rejected in the end. This was confusing
me.
This adds that by emitting a dump for the failed case.
Note the patch is little more complex as we don't want to print out
if debug counter rejected the pattern and then we need to fix up
when we mark needing a label or not.

OK? Bootstrapped and tested on x86_64-linux-gnu with no regressions.

gcc/ChangeLog:

	* genmatch.cc (needs_label): New variable
	(expr::gen_transform): Set needs_label
	if we use the local_label.
	(dt_simplify::gen_1): Use `_1` for the debug count label.
	After the local label, emit debug print for the failure.
	Emit `_1` label if needed.