ada: Reduce span of variable
This patch does not change the behavior of the compiler, but is
intended to improve readability. It seizes an opportunity to move
a variable declaration to a smaller scope, so that it's clearer
that the variable is not used outside of that scope.

gcc/ada/

	* sem_ch3.adb (Replace_Type): Reduce span of variable.

Tested on x86_64-pc-linux-gnu, committed on master.