ada: Simplify copying of node lists
When creating a copy of a node list we called Copy_Entity for entities
and Copy_Separate_Tree for other nodes. This was unnecessary, because
the Copy_Separate_Tree when called on entities will just do Copy_Entity.

Code cleanup; semantics is unaffected.

gcc/ada/

	* atree.adb (Copy_List): Call Copy_Separate_Tree for both entities and
	other nodes.

Tested on x86_64-pc-linux-gnu, committed on master.