ada: Deconstruct a no longer used parameter of New_Copy_Tree
Parameter Scopes_In_EWA_OK of New_Copy_Tree was introduced in 2018 to
deal with expressions-with-actions (EWA) in the build-in-place
machinery. However, after changes made in 2022 it is no longer used by
any caller.

Cleanup related to handling of expression functions in GNATprove;
semantics is unaffected.

gcc/ada/

	* sem_util.ads (New_Copy_Tree): Remove Scopes_In_EWA_OK from spec;
	adapt comment.
	* sem_util.adb (New_Copy_Tree): Remove Scopes_In_EWA_OK from body;
	adapt code.

Tested on x86_64-pc-linux-gnu, committed on master.