ada: Require successful build of xsnamest
While experimenting, I introduced a compilation error into xsnamest.
This took a little while to track down because, while the error was in
the log, the build did not stop.  This patch changes Make-generated.in
to require a successful build of this program.

gcc/ada/

	* Make-generated.in (ada/stamp-snames): Check result of
	gnatmake.

Tested on x86_64-pc-linux-gnu, committed on master.