ada: Missing warning on null-excluding array aggregate component
The compiler does not report warnings on the initialization
of arrays of null-excluding access type components by means
of iterated component association, when the expression
initializing each component is either a conditional
expression or a case expression that may initialize
some component with a null value.

gcc/ada/

	* sem_aggr.adb
	(Warn_On_Null_Component_Association): New subprogram.
	(Empty_Range): Adding missing support for iterated component
	association node.
	(Resolve_Array_Aggregate): Report warning on iterated component
	association that may initialize some component of an array of
	null-excluding access type components with a null value.
	* exp_ch4.adb
	(Expand_N_Expression_With_Actions): Add missing type check since
	the subtype of the EWA node and the subtype of the expression
	may differ.

Tested on x86_64-pc-linux-gnu, committed on master.