tree-optimization/109791 - expand &x + off for niter compute
The following makes expand_simple_operations expand POINTER_PLUS_EXPRs
with variable offset when the base is invariant.  That will allow
to simplify address differences to offset differences in some cases.
Note the patch doesn't follow the variable off chain as I don't have
a testcase showing that's beneficial.

Bootstrapped and tested on x86_64-unknown-linux-gnu.

I'm going to push this when the last patch in the series tests OK
and the testcase from PR109791 then can be added.

	PR tree-optimization/109791
	* tree-ssa-loop-niter.cc (expand_simple_operations): Expand &x + off.